refactor: 기존 이벤트 객체 제거 및 common 모듈의 이벤트 객체로 연결
