--- src/main/java/com/yeoljeong/tripmate/domain/events/PlanUnitConfirmed.java
@@ -1,13 +1,0 @@
-package com.yeoljeong.tripmate.domain.events;
-
-import java.util.List;
-import java.util.UUID;
-
-public record PlanUnitConfirmed(
-    String eventHash,
-    UUID planUnitId,
-    String planTitle,
-    List<UUID> receivers
-) {
-
-}
